requirements.txt
langchain==0.2.10
langchain-openai==0.1.17
python-dotenv==1.0.1
langchain-groq==0.1.6
langchain-community== 0.2.10
langchain-chroma>=0.1.2
langsmith
bs4==0.0.2
unstructured==0.15.0
fastapi
uvicorn